requirements.txt

Django==           2.1.7  
django-authtools== 1.6.0  
Pillow==           5.4.1  
pip     ==         19.0.3 
psycopg2  ==       2.7.7  
pytz        ==     2018.9 
setuptools    ==   40.8.0 
wheel          ==  0.33.1 
